created seperate files for types and constants
