+++ src/app/dashboard/home/page.tsx
@@ -2,13 +2,11 @@ import {
     Card,
     CardContent,
     CardDescription,
-    CardFooter,
     CardHeader,
     CardTitle,
   } from "@/components/ui/card"
 import { Table,
     TableBody,
-    TableCaption,
     TableCell,
     TableHead,
     TableHeader,
@@ -16,94 +14,7 @@ import { Table,
 
 import Link from 'next/link'
 import OngoingProjectCard from "./OngoingProjectCard";
-
-
-interface project{
-    id:number
-    name: string;
-    start_date?:string;
-    end_date?:string;
-    budget?:string;
-    location?:string;
-    status:boolean;
-} 
-
-type projectArray = project[];
-
-const projects:projectArray = [
-    {
-        id: 1,
-        name:"Bruce Lee:The Man and the Legend",
-        start_date:"",
-        end_date:"",
-        budget:"",
-        location:"",
-        status:false,
-    },
-    {
-        id:2, 
-        name:"Til Madness Do Us Part",
-        start_date:"",
-        end_date:"",
-        budget:"",
-        location:"",
-        status: false,
-    },
-    {
-        id: 3, 
-        name:"Sakura Dreams: A Journey Through Japan's Contrasting Culture",
-        start_date:"",
-        end_date:"",
-        budget:"",
-        location:"",
-        status:false,
-    },
-    {
-        id: 4, 
-        name: "The Secret Garden",
-        start_date: "2024-06-01",
-        end_date: "2024-09-30",
-        budget: "$50,000,000",
-        location: "England",
-        status: true
-    },
-    {
-        id: 5,
-        name: "Inception",
-        start_date: "2024-03-15",
-        end_date: "2024-08-30",
-        budget: "$160,000,000",
-        location: "United States",
-        status: true
-    },
-    {
-        id: 6,
-        name: "Avatar 2",
-        start_date: "2024-10-01",
-        end_date: "2025-03-31",
-        budget: "$300,000,000",
-        location: "New Zealand",
-        status: true
-    },
-    {
-        id: 7,
-        name: "Jurassic World: Dominion",
-        start_date: "2023-07-01",
-        end_date: "2024-05-30",
-        budget: "$200,000,000",
-        location: "United Kingdom",
-        status: true
-    },
-    {
-        id: 8,
-        name: "Wonder Woman 3",
-        start_date: "2025-01-15",
-        end_date: "2025-06-30",
-        budget: "$120,000,000",
-        location: "United States",
-        status: true
-    }    
-]
+import { projects } from "@/constants";
 
 const page = () => {
