+++ src/components/sidebar/components/SideBarButton.tsx
@@ -8,13 +8,13 @@ import { useSideBarControl } from '@/contexts/SideBarContext'
 import { useProjectControl } from '@/contexts/ProjectContext'
 
 
-const SideBarButton = ({icon, text, link, root}:{
-    icon: React.ReactNode,
+const SideBarButton = ({Icon, text, link, root}:{
+    Icon: React.FC,
     text: string,
     link: string, 
     root: string
 }) => {
-    const {toggle} = useSideBarControl()
+    const {setisSideBarOpen} = useSideBarControl()
     const segments = useSelectedLayoutSegments()
     const length = segments.length
     const {setProject} = useProjectControl()
@@ -29,17 +29,14 @@ const SideBarButton = ({icon, text, link, root}:{
         if(root==='dashboard'){
             setProject({id:"", name:""})
         }
-        toggle
-        
+        setisSideBarOpen(false)
     }
   return (
     <Link className={`${segments[length-1]==link ? 'active' : ''}`} href={url}>
         <Button onClick={handleClick} variant="ghost"
             className={`${segments[length-1]==link ? 'bg-gradient-to-tr from-gray-900 to-gray-800 text-white hover:text-white shadow-md shadow-gray-900/10 hover:shadow-lg hover:shadow-gray-900/20 active:opacity-[0.85]' 
             : 'text-[#607D8B] hover:bg-[#607D8B]/10 active:bg-[#607D8B]/30'} w-full h-auto flex items-center gap-4 px-4 py-3 capitalize justify-start`}>
-            {
-                icon
-            }
+            <Icon />
             <p className="block font-sans antialiased text-base leading-relaxed text-inherit font-medium capitalize">
                 {text}
             </p>
